Implementa novos testes e corrige antigos
